pyproject.toml
[project]
name = "rag_test_suite"
version = "0.1.0"
description = "Automated testing framework for RAG-based CrewAI chat crews"
authors = [{ name = "Konecta", email = "tech@konecta.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[litellm,tools]==1.8.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
rag_test_suite = "rag_test_suite.main:main"
kickoff = "rag_test_suite.main:run_flow_entry"
run_crew = "rag_test_suite.main:run_flow_entry"
run_flow = "rag_test_suite.main:run_flow_entry"
run_with_trigger = "rag_test_suite.main:run_flow_with_trigger"
train = "rag_test_suite.main:train"
replay = "rag_test_suite.main:replay"
test = "rag_test_suite.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_test_suite"]

[tool.crewai]
type = "flow"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning",
    "ignore::pytest.PytestReturnNotNoneWarning",
]
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (may require external services)",
    "slow: Slow tests (skip with -m 'not slow')",
    "requires_env: Tests requiring environment variables",
]
